Use importlib.machinery.ModuleSpec everywhere
